(PC-5884) Move SignIn from context to the api file
