+++ src/features/auth/AuthContext.tsx
@@ -2,21 +2,21 @@ import React, { useContext, useEffect, useState } from 'react'
 import { useQueryClient } from 'react-query'
 
 import { api } from 'api/api'
-import { AccountRequest, SigninRequest, SigninResponse } from 'api/gen'
+import { AccountRequest, SigninResponse } from 'api/gen'
 import { firebaseAnalytics } from 'libs/analytics'
 import { clearRefreshToken, saveRefreshToken } from 'libs/keychain'
 import { clearAccessToken, getAccessToken, saveAccessToken } from 'libs/storage'
 
 export interface IAuthContext {
   isLoggedIn: boolean
-  signIn: (data: SigninRequest) => Promise<boolean>
+  setIsLoggedIn: (isLoggedIn: boolean) => void
   signUp: (data: AccountRequest) => Promise<boolean>
   signOut: () => Promise<void>
 }
 
 export const AuthContext = React.createContext<IAuthContext>({
   isLoggedIn: false,
-  signIn: () => Promise.resolve(false),
+  setIsLoggedIn: () => void 0,
   signUp: () => Promise.resolve(false),
   signOut: () => Promise.resolve(),
 })
@@ -48,19 +48,6 @@ export const AuthWrapper = ({ children }: { children: Element }) => {
     })
   }, [])
 
-  const signIn = async (body: SigninRequest) => {
-    try {
-      const response = await api.postnativev1signin(body, { credentials: 'omit' })
-      if (!response) return false
-
-      await loginRoutine(response, 'fromLogin')
-      setIsLoggedIn(true)
-      return true
-    } catch (error) {
-      return false
-    }
-  }
-
   const signUp = async (body: AccountRequest) => {
     try {
       const response = await api.postnativev1account(body, { credentials: 'omit' })
@@ -78,15 +65,10 @@ export const AuthWrapper = ({ children }: { children: Element }) => {
   }
 
   return (
-    <AuthContext.Provider value={{ isLoggedIn, signIn, signUp, signOut }}>
+    <AuthContext.Provider value={{ isLoggedIn, setIsLoggedIn, signUp, signOut }}>
       {children}
     </AuthContext.Provider>
   )
-}
-
-export function useSignIn(): IAuthContext['signIn'] {
-  const authContext = useAuthContext()
-  return authContext.signIn
 }
 
 export function useSignOut(): IAuthContext['signOut'] {
