+++ src/features/auth/__mocks__/api.ts
@@ -1,0 +1,3 @@
+export const signIn = jest.fn()
+
+export const useSignIn = jest.fn(() => signIn)
